+++ Makefile
@@ -4,5 +4,5 @@ composer:
 	composer install
 
 test: composer
-	vendor/bin/phpunit tests --colors --debug --verbose
+	vendor/bin/phpunit test --colors --debug --verbose
