+++ test/Unit/Serializer/Transformer/Entity/TagTransformerTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace RCatlin\Blog\Tests\Serializer\Transformer\Entity;
+namespace RCatlin\Blog\Test\Unit\Serializer\Transformer\Entity;
 
 use RCatlin\Blog\Entity;
 use RCatlin\Blog\Serializer\Transformer\Entity\TagTransformer;
-use RCatlin\Blog\Tests\HasFaker;
+use RCatlin\Blog\Test\Unit\HasFaker;
 
 class TagTransformerTest extends \PHPUnit_Framework_TestCase
 {
