+++ test/Unit/Validator/Rule/IsArrayRuleTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace RCatlin\Blog\Tests\Validator\Rule;
+namespace RCatlin\Blog\Test\Unit\Validator\Rule;
 
-use RCatlin\Blog\Tests\HasFaker;
+use RCatlin\Blog\Test\Unit\HasFaker;
 use RCatlin\Blog\Validator\Rule;
 
 class IsArrayRuleTest extends \PHPUnit_Framework_TestCase
