--- tests/FooTest.php
@@ -1,10 +1,0 @@
-<?php
-
-
-class FooTest extends \PHPUnit_Framework_TestCase
-{
-    public function testAssertion()
-    {
-        $this->assertTrue(true);
-    }
-}
